x/leaderboard/genesis.go
package leaderboard

import (
	sdk "github.com/cosmos/cosmos-sdk/types"
	"github.com/perfogic/cosmos-checkers/x/leaderboard/keeper"
	"github.com/perfogic/cosmos-checkers/x/leaderboard/types"
)

// InitGenesis initializes the capability module's state from a provided genesis
// state.
func InitGenesis(ctx sdk.Context, k keeper.Keeper, genState types.GenesisState) {
	// Set if defined
	k.SetLeaderboard(ctx, genState.Leaderboard)
	// this line is used by starport scaffolding # genesis/module/init
	k.SetParams(ctx, genState.Params)
}

// ExportGenesis returns the capability module's exported genesis.
func ExportGenesis(ctx sdk.Context, k keeper.Keeper) *types.GenesisState {
	genesis := types.DefaultGenesis()
	genesis.Params = k.GetParams(ctx)

	// Get all leaderboard
	genesis.Leaderboard = k.GetLeaderboard(ctx)
	// this line is used by starport scaffolding # genesis/module/export

	return genesis
}
